+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project are documented in this file.
 
 ## [Unreleased]
 
+### Added
+- Added a minimap button that opens the in-game options window on left click.
+
 ## [1.0.15-beta.2] - 2026-03-22
 
 ### Added